Also move PythonSetup/PythonRepos subsystems out of the python backend.
This is a longstanding TODO: Pants needs these for plugin loading even
if the repo contains no python code at all. So these belong in core
pants, not in the python backend.

Note that these are not public API, and their options scopes
haven't changed, so this move does not require deprecation.
