+++ src/python/pants/backend/python/tasks/BUILD
@@ -17,7 +17,6 @@ python_library(
     'src/python/pants/backend/python:interpreter_cache',
     'src/python/pants/backend/python:python_chroot',
     'src/python/pants/backend/python:python_requirement',
-    'src/python/pants/backend/python:python_setup',
     'src/python/pants/backend/python:thrift_builder',
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/base:build_environment',
@@ -31,6 +30,7 @@ python_library(
     'src/python/pants/build_graph',
     'src/python/pants/ivy',
     'src/python/pants/option',
+    'src/python/pants/python',
     'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
