+++ src/python/pants/python/BUILD
@@ -1,0 +1,10 @@
+# Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_library(
+  dependencies = [
+    '3rdparty/python:pex',
+    '3rdparty/python:setuptools',
+    'src/python/pants/subsystem',
+  ]
+)
