+++ tests/python/pants_test/backend/python/BUILD
@@ -24,11 +24,11 @@ python_tests(
     'src/python/pants/backend/python:interpreter_cache',
     'src/python/pants/backend/python:python_chroot',
     'src/python/pants/backend/python:python_requirement',
-    'src/python/pants/backend/python:python_setup',
     'src/python/pants/binaries:binary_util',
     'src/python/pants/binaries:thrift_util',
     'src/python/pants/ivy',
     'src/python/pants/java/distribution',
+    'src/python/pants/python',
     'src/python/pants/util:contextutil',
     'tests/python/pants_test/subsystem:subsystem_utils',
     'tests/python/pants_test:base_test',
