+++ tests/python/pants_test/backend/python/tasks2/BUILD
@@ -12,7 +12,6 @@ python_tests(
     'src/python/pants/backend/python:interpreter_cache',
     'src/python/pants/backend/python:python_artifact',
     'src/python/pants/backend/python:python_requirement',
-    'src/python/pants/backend/python:python_setup',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks2',
     'src/python/pants/base:build_root',
@@ -20,6 +19,7 @@ python_tests(
     'src/python/pants/base:run_info',
     'src/python/pants/build_graph',
     'src/python/pants/fs',
+    'src/python/pants/python',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test/tasks:task_test_base',
