+++ tests/python/pants_test/backend/python/tasks2/test_gather_sources.py
@@ -10,10 +10,10 @@ import os
 from pex.interpreter import PythonInterpreter
 
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
-from pants.backend.python.python_setup import PythonRepos, PythonSetup
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.tasks2.gather_sources import GatherSources
 from pants.build_graph.resources import Resources
+from pants.python.python_setup import PythonRepos, PythonSetup
 from pants_test.tasks.task_test_base import TaskTestBase
